+++ src/WebResponses.cpp
@@ -23,7 +23,7 @@
 #include "cbuf.h"
 
 #ifdef ASYNCWEBSERVER_DEBUG_TRACE
-#define DEBUG_PRINTFP(fmt, ...) Serial.printf_P(PSTR("[%d]" fmt), millis(), ##__VA_ARGS__)
+#define DEBUG_PRINTFP(fmt, ...) Serial.printf_P(PSTR("[%u]" fmt), (unsigned) millis(), ##__VA_ARGS__)
 #else
 #define DEBUG_PRINTFP(...)
 #endif
